requirements.txt
django-compressor
dj-database-url
gunicorn
django-blog-zinnia
South
Django>=1.6
